Add files via upload
The Love-O-Meter is a project featuring the use of a temperature sensor. Using a formula to convert voltage into temperature, the temperature sensor will be able to detect any increase in heat. The bigger the margin between baseline temperature and current temperature, the more LED lights will switch on the breadboard.