Add files via upload
The Love-O-Meter is a project featuring the use of a temperature sensor. Utilizing some equations, voltage is converted to temperature to allow the temperature sensor to achieve a baseline temperature. Any additional change to the temperature will result on the corresponding LEDs to light up depending on the margin between baseline temperature and the temperature recorded in the instance.